# Expose more tessellator method