* Make public the Tessellator methods to tessellate a circle, a
    mesh, a rectangle, a line, a path, a quadratic and cubic
    bezier curve.
* Add doc to tessellate_text & make it take a reference.
* Add Mesh::append_ref method.

## Notes
* I'm not too happy with the `Mesh::append_ref` method name, but I couldn't think of anything else.
* I'm not exactly sure what should be the behavior of `Mesh::append_ref`
* This is a breaking change because tessellate_text now takes a reference. A non-breaking alternative would be to make it take a `AsRef<TextShape>`.

Closes #1383