@jean-airoldie 2022-03-20
A side effect of not having to clone is that it improves the `egui_demo_lib::tessellate_text benchmark`.
## before
```
tessellate_text         time:   [68.282 ns 69.360 ns 70.408 ns]
                        change: [+83.723% +86.104% +88.444%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

## after
```
tessellate_text         time:   [36.897 ns 37.308 ns 37.777 ns]
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe
```